requirements.txt
streamlit>=1.40.2
streamlit-cookies-manager==0.2.0
streamlit-option-menu==0.4.0
argon2-cffi==23.1.0
pandas~=2.2.3
beautifulsoup4==4.11.1
requests~=2.27.1
mysql-connector-python~=9.1.0
Pillow~=9.1.1
python-dotenv==0.20.0
selenium~=4.18.1
chromedriver-autoinstaller~=0.6.4
lxml==5.3.0